refactor(core): Use a rest controller for status endpoints (#388)
* test(core): Update tests to test statusHandler

The status endpoint tests are currently testing AllStatusEndpoint and
InstanceStatusEndpoint which are pass-through classes that will be
removed; update the tests to directly test StatusHandler.

* refactor(core): Use a rest controller for status endpoints

The Endpoint classes that are used to implement the status endpoint
have significantly changed in Spring Boot 2. Instead of creating and
exposing an endpoint, just use a rest controller.

* fix(core): Remove package from componentscan

The prior commit moved everything out of the endpoints package into
other packages, so we should remove it from the component scan.
